Camera + Dragon Assets
Added a new dragon asset into the scene.

Added Cinemachine into the project and connected it to follow the dragon.

Created a character controller to move the dragon around the scene.

Placed the mastercanvas on top of the dragon in world space and made it always look at the camera.